move scheduler to separate folder
